11.코드정리.
