+++ src/main/java/com/ejoongseok/wmslive/outbound/feature/RegisterOutbound.java
@@ -6,9 +6,7 @@ import com.ejoongseok.wmslive.outbound.domain.Order;
 import com.ejoongseok.wmslive.outbound.domain.OrderProduct;
 import com.ejoongseok.wmslive.outbound.domain.OrderRepository;
 import com.ejoongseok.wmslive.outbound.domain.Outbound;
-import com.ejoongseok.wmslive.outbound.domain.OutboundProduct;
 import com.ejoongseok.wmslive.outbound.domain.OutboundRepository;
-import com.ejoongseok.wmslive.outbound.domain.PackagingMaterial;
 import com.ejoongseok.wmslive.outbound.domain.PackagingMaterialRepository;
 import com.ejoongseok.wmslive.outbound.domain.PackagingMaterials;
 import jakarta.validation.Valid;
@@ -31,7 +29,7 @@ public class RegisterOutbound {
     private final OutboundRepository outboundRepository;
     private final InventoryRepository inventoryRepository;
     private final PackagingMaterialRepository packagingMaterialRepository;
-    private final ConsturctOutbound consturctOutbound = new ConsturctOutbound();
+    private final ConstructOutbound constructOutbound = new ConstructOutbound();
 
 
     @PostMapping("/outbounds")
@@ -43,7 +41,7 @@ public class RegisterOutbound {
         final List<Inventories> inventoriesList = inventoriesList(order.orderProducts());
         final PackagingMaterials packagingMaterials = new PackagingMaterials(packagingMaterialRepository.findAll());
 
-        final Outbound outbound = consturctOutbound.create(
+        final Outbound outbound = constructOutbound.create(
                 inventoriesList,
                 packagingMaterials,
                 order,
@@ -68,32 +66,7 @@ public class RegisterOutbound {
             final Order order,
             final Boolean isPriorityDelivery,
             final LocalDate desiredDeliveryAt) {
-        return consturctOutbound.create(inventoriesList, packagingMaterials, order, isPriorityDelivery, desiredDeliveryAt);
-    }
-
-    private void validateInventory(final List<Inventories> inventoriesList, final List<OrderProduct> orderProducts) {
-        consturctOutbound.validateInventory(inventoriesList, orderProducts);
-    }
-
-    private Inventories getInventories(final List<Inventories> inventoriesList, final OrderProduct orderProduct) {
-        return consturctOutbound.getInventories(inventoriesList, orderProduct);
-    }
-
-    private Outbound newOutbound(
-            final Order order,
-            final PackagingMaterial packagingMaterial,
-            final Boolean isPriorityDelivery,
-            final LocalDate desiredDeliveryAt) {
-        return consturctOutbound.newOutbound(order, packagingMaterial, isPriorityDelivery, desiredDeliveryAt);
-    }
-
-    private List<OutboundProduct> mapToOutboundProducts(
-            final List<OrderProduct> orderProducts) {
-        return consturctOutbound.mapToOutboundProducts(orderProducts);
-    }
-
-    private OutboundProduct newOutboundProduct(final OrderProduct orderProduct) {
-        return consturctOutbound.newOutboundProduct(orderProduct);
+        return constructOutbound.create(inventoriesList, packagingMaterials, order, isPriorityDelivery, desiredDeliveryAt);
     }
 
     public record Request(
